Substance
Inspired by the movie "Substance". Imagine we have a substance that allows us to create a "better" version of a living creature. In our case, applying the substance will only transform a living creature into a new type, halving the current number of days lived (rounded up). Keep in mind that in comparison with the movie, the second creature is not created. The number of days lived (daysLived) for each animal is limited to 10 days – the animal dies as soon as the 11th day arrives. It is recommended to apply the substance to the animal only once, but if the substance is overused, the "better" version of the animal will turn into a monster that automatically kills all nearby animals on the same day.

The substance can be removed from the organisms of only the "better" types of the animals, then the animal will revert to a normal type, and the days lived will be doubled. If in this case the number of days lived exceeds 10, the animal will die immediately upon the arrival of the next day.

The types of animals are shown in the class diagram. The substance can be applied to a creature of type Mouse/Fish/Bird, and then this creature will become a "better" type: BetterMouse/BetterFish/BetterBird. When the substance is applied to an animal of a "better" type, this animal becomes a Monster, which is irrevertible type of creature living 1 day (his daysLived at most 1). Any living creature can attack any other creature – in this case, the one who is attacked dies.

All animals are strictly distributed according to their type in template-based containers: Cage<T>, Aquarium<T>. The animals are placed in containers in increasing order of their days lived (for animals with the same number of days lived use the lexicographical order of names). Some animals are luckier and are free – they are in container Freedom<A>, which is suitable for any type of animal. In Freedom<A> any animals, even monsters, cannot kill anyone, also we cannot apply or remove the substance while the animal is free. A cage can be either for mice or for birds, but if you try to put a fish there, you won't succeed. Similarly, with an aquarium, it can be either for mice or for fish, but you cannot put a bird there. "Better" types of animals have their own separate containers: Fish, BetterFish, Mouse, BetterMouse, Bird and BetterBird are in different containers. Applying or removing the substance for the animal moves this animal to another type of container suitable for it. When a Monster appears in a container, all animals in this container die, and the Monster is then moved to freedom. For killing of other animals Monster does not require any specific command. Create containers of all possible animal types except Monsters in advance: Cage<Bird>, Cage<BetterBird>, Cage<Mouse>, Cage<BetterMouse>, Aquarium<Fish>, Aquarium<BetterFish>, Aquarium<Mouse>, Aquarium<BetterMouse>, Freedom<Animal>- use this order when updating daysLived. Reorder animals in the container on every update.

Animals can only attack each other within the same container.