Cards
 Ahmed and Karim are playing a game with cards.
    Each card has a positive digit X written on it.
    Ahmed has a collection of N cards, and Karim has a collection of M cards.
    Neither of them know which cards the other one has. 
They independently reorder their collections of cards for some time.
Finally, the chosen arrangements are presented to you (the judge) to determine the winner.
    You — being a smart judge — decided to write a program to calculate the scores and determine the winner. 
The game scoring system is defined as follows:
    Cards are considered from left to right, one by one.
    If a card with the value X is present in one collection only but not the other, the collection owner does not gain any points for it.
    If a card with the value X is present in both collections (i.e., at positions i
in Ahmed's collection and position j in Karim's collection), then the earliest occurrences of the card in each collection are considered in a duel.
    If i<j, Ahmed gets a point. Similarly, if j<i, Karim gets a point.
If i=j, no one gets any points.
        The cards used during the duel are then flipped and can never be reconsidered. 
In the end, the player with more points wins. A tie is possible if Ahmed and Karim have the same number of points.